chore(rust): clippy 113 → 62 — `--fix` 를 막고 있던 두 자리를 풀었다
A-6 껍질의 나머지 절반. **0 은 아직 못 만들었다** — 남은 62 는 손으로 봐야 한다.

🔴 **`cargo clippy --fix` 가 한 건도 안 먹고 있었다.** 「38건을 고칠 수 있다」고
말만 하고 파일이 하나도 안 바뀌기에 로그를 끝까지 읽으니, 제안 둘이 `E0502`
(빌림 충돌)를 내고 **한 자리가 깨지면 cargo 가 그 배치의 수정을 전부 되돌린다.**
도구가 「안 된다」고 말한 게 아니라 **조용히 되돌리고 있었다** — 이 저장소가
반복해 본 형태다.

막고 있던 둘에 `#[allow]` + WHY 를 달자 나머지 40건이 한 번에 들어갔다:
- `npc_sim.advance_all_grades` — `is_some()`+`unwrap()` 을 `if let` 으로 바꾸면
  그 뒤 `npc` 의 mutable 빌림과 충돌한다. `grade` 는 `Option<u8>`(Copy)라
  지금 꼴이 맞다
- `npc_sim.run_offseason` — `origin_league.to_string()` 의 `.to_string()` 이
  바로 `npc` 빌림을 끊는 자리다. 빼면 `fa_fallback(npc, …)` 와 충돌한다

들어간 40건은 전부 **뜻이 같은 치환**이다(`map_or(true,…)`→`is_none_or` ·
`% 4 == 0`→`is_multiple_of(4)` · 불필요 캐스트·수명·`into_iter()` 제거 ·
`impl Default` 를 test module 앞으로). **숫자 상수는 한 자리도 안 바뀌었다.**

Rust 글자를 읽던 TS 검사 넷도 새 꼴로 고쳤다(faUnsignedFate · proAllstar ·
defenseRecord · pinchHit).

**남은 62 — 왜 안 했나**(내역은 `lib.rs` 머리말):
- 38 `assertions_on_constants` — 검사 안의 `assert!` 가 상수로 접힌다.
  한 건씩 「무엇을 지키던 검사였나」를 봐야 한다
- 9 `manual_clamp` — 🔴 **밸런스 식이다**(OVR·연봉 산식). `.max().min()` 과
  `clamp` 는 NaN 에서 갈리고 TS 검사가 그 글자를 못박는다. **밸런스 동결 중이라
  손대지 않았다**
- 8 `too_many_arguments` — `#[allow]` 를 달지 서명을 쪼갤지 판단이 필요하다
- 7 그 밖

⚠ **`warn` → `deny` 와 CI `-D warnings` 는 안 건드렸다.** 0 이 아닌 채로 걸면
CI 가 그날부터 빨강이다(그 주석이 처음부터 그렇게 적혀 있다). 주석만 지금 수로
갱신했다(`lib.rs` · 루트 `.github/workflows/ci.yml`).

회귀: cargo test 365 초록 · `build:native` 도장 `78e07bdcef08` · `check:native` OK ·
vitest 2816 초록 · tsc 0 · eslint 오류 0 · `format:check` 전량 초록 ·
🔴 **`check:measurerepro` 2회 동일이고 고치기 전과 같은 값**
(대학합격3 · 3년차 OVR 70 · 구속 72 · 완주) — 동작이 안 바뀐 증거다

+++ 02.SvelteElectron/packages/engine-native/src/team_engine.rs
@@ -1229,8 +1229,7 @@ const TRADE_FLEX_FLOOR: f64 = 0.05;
 
 pub fn eval_trade_value(p: EvalTradeValueParams) -> TradeEvalResult {
     let profile = &p.team_profile;
-    let flex = (((p.salary_cap - p.current_payroll) as f64 / p.salary_cap as f64)
-        as f64).max(TRADE_FLEX_FLOOR);
+    let flex = ((p.salary_cap - p.current_payroll) as f64 / p.salary_cap as f64).max(TRADE_FLEX_FLOOR);
 
     let value_asset = |asset: &TradeAsset| -> f64 {
         let mut v = asset.ovr * 1.5;
